+++ SDM/Models/ReportModels/FullDatabaseModel.cs
@@ -11,7 +11,7 @@ namespace SDM.Models.ReportModels
     {
         public string ClientId { get; set; }
         public int InvoiceNumber { get; set; }
-        public int PaymentDue { get; set; }
+        public float PaymentDue { get; set; }
         public DateTime PaymentDueDate { get; set; }
         public List<PaymentDateLatencyPaid> Payments { get; set; } = new List<PaymentDateLatencyPaid>();
     }
@@ -19,7 +19,7 @@ namespace SDM.Models.ReportModels
     public class PaymentDateLatencyPaid
     {
         public DateTime PaymentDate { get; set; }
-        public int PaymentPaid { get; set; }
+        public float PaymentPaid { get; set; }
         public int Latency { get; set; }
     }
 }
